requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# NLP & ML
transformers>=4.36.0
sentence-transformers>=2.2.0
torch>=2.1.0
spacy>=3.7.0
scikit-learn>=1.3.0

# Data ingestion
requests>=2.31.0
beautifulsoup4>=4.12.0
pymupdf>=1.23.0
lxml>=5.0.0

# Vector search
faiss-cpu>=1.7.4

# Agent
langchain>=0.1.0
langchain-community>=0.0.20

# MLOps
mlflow>=2.10.0
optuna>=3.4.0

# Frontend
streamlit>=1.29.0

# PDF report
reportlab>=4.0.0
